:recycle: move exceptions to core package
