+++ caikit/core/data_model/streams/converter.py
@@ -26,7 +26,7 @@ from typing import Callable, Dict, List
 import alog
 
 # Local
-from ...toolkit.errors import error_handler
+from ...exceptions import error_handler
 from .. import DataStream
 
 log = alog.use_channel("DATSTRMCNVRTR")
