+++ caikit/core/exceptions/__init__.py
@@ -12,5 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 # Local
+from .validation_error import DataValidationError
