+++ caikit/core/modules/config.py
@@ -21,7 +21,7 @@ import alog
 
 # Local
 from .. import toolkit
-from ..toolkit.errors import error_handler
+from ..exceptions import error_handler
 
 log = alog.use_channel("MODULE_CFG")
 error = error_handler.get(log)
