+++ caikit/core/toolkit/factory.py
@@ -25,7 +25,7 @@ import aconfig
 import alog
 
 # Local
-from .errors import error_handler
+from ..exceptions import error_handler
 
 log = alog.use_channel("FCTRY")
 error = error_handler.get(log)
