+++ caikit/core/toolkit/quality_evaluation.py
@@ -26,7 +26,7 @@ import re
 import alog
 
 # Local
-from .errors import error_handler
+from ..exceptions import error_handler
 
 log = alog.use_channel("TLKIT")
 error = error_handler.get(log)
