+++ caikit/core/toolkit/wip_decorator.py
@@ -22,7 +22,7 @@ import inspect
 import alog
 
 # Local
-from .errors import error_handler
+from ..exceptions import error_handler
 
 log = alog.use_channel("WIPDC")
 error = error_handler.get(log)
