+++ tests/core/data_model/streams/test_validator.py
@@ -15,7 +15,7 @@
 # Local
 from caikit.core.data_model import DataStream
 from caikit.core.data_model.streams.validator import DataStreamValidator
-from caikit.core.toolkit.errors import DataValidationError
+from caikit.core.exceptions.validation_error import DataValidationError
 
 # Unit Test Infrastructure
 from tests.base import TestCaseBase
